requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.9.0
langchain-core==0.2.2
langchain-community==0.1.0
langchain==0.1.0
boto3==1.18.67
python-dotenv==0.19.2
